examples/dashboard-operator: fake vendor directory
Fake the vendoring of the kubebuilder-declarative-operator library by
copying it into vendor in just the right spot.

This odd move is needed to builder the Docker container. I'd like to
make it look as much like the walkthrough as possible, or else I'd
modify the Dockerfile to live at the root of the project.