declarative,docs: remove packageName
The addon manifest loader uses the 'ComponentName()' call on
'CommonSpec' to resolve the component to load the manifest from.

We could do this directly in declarative by requiring packageName and
bubbling it out to the Loader, but there's no reason for the declarative
controller to keep track of this value. The library can't make decisions
about which package to use beyond passing a string it was given.

Pushing this out to the addon/loader allows implementers to own the
decision and perhaps change it at runtime based on the CRs configuration
or the state of CLI flags.

fixes #33