addon, declarative: add Patch
Introduce Patch. An Addon can implement the Patchable interface to
accept a set of Patches that are overlaid on the deployment. For
example:

apiVersion: addons.sigs.k8s.io/v1alpha1
kind: Dashboard
metadata:
  name: dashboard-sample
  namespace: kube-system
spec:
  patches:
  - apiVersion: apps/v1beta2
    kind: Deployment
    metadata:
      name: kubernetes-dashboard
      namespace: kube-system
    spec:
      replicas: 5

This patch will change the replica count of the dashboard deployment.
This allows us to provide a break-glass experience for the user to set
configuration options while not requiring every configuration option be
directly exposed on the CRD.

- It's not clear if we want this on all addons. If so, we should
  graduate it to CommonObject
- Patch doesn't work if it happens after the addons.TransformApplicationFromStatus
  transform. It raises a DeepCopy error. This is a bug.
- This patching is equivalent/borrowed from kustomize which is not
  currently avaliable as a library.