Add metrics-server-operator
This commit adds metrics-server-operator to `examples/`.
As of now dashboard-operator is located in `examples/`.
Dashboard v2.0.0 which is next relase can use metrics-server for
monitoring statistics of k8s nodes.

It's better to add metrics-server-operator to this repo because that
dashboard v2.0.0 will release before long.

When dashboard v2.0.0 has released, it is needed to add v2.0.0
manifests to dashboard-operator.