Don't bail out if applying CRD + CR in manifest
The kubectl library upgrade for 1.18 now checks if all types are
available in applied manifests in res.Infos(). Callers should try to
apply everything returned if non-nil rather than immediately exiting on
an error.

This CL attempts to apply even if res.Infos() encounters an error. If
both .Apply() and res.Infos() return an error, returns the error from
.Apply().