Add side input for golden file tests
Adds a posibility to provide a *.side_in.yaml file, that contains
resources that should exist before the resource defined in `*.in.yaml` is applied.
These side input resources are going to be created before running a testcase, and
cleaned up at the end of it. The resources will be accessible via Get and List.
Note that the existing functionality has not been changed.