testreconciler: pause after reconciliation
We sleep briefly (100ms) after updating the status.

This is a hack to ensure that we see the watch event, because
otherwise we can start a re-reconcile based on our derived objects,
and then reconcile again based on our own status update.

This is racy behaviour and causes non-determinism which breaks our
golden tests.

We do this only in this test controller, but maybe we should have
similar logic to "debounce" in all controllers.