feat: nested manifests in object transforms
This change adds support for applying object transforms to nested
manifests. The nested manifest pattern is a pattern in which an object
has a string field which consists of its own yaml manifest.

As an example, a ConfigMap may contain a manifest bundle in one of its
data fields.

A nested manifest will recurse and apply all object transforms to the
underlying objects and then write the result back to the parent object.